Refactored cucumber steps
