+++ app/views/calendar/calendar_template.html.slim
@@ -26,7 +26,7 @@ div.row
 
 div#autoscheduler ng-if="schedulerReady && isOwner"
   div.page-header
-    h2 Auto-scheduler
+    h2#autoscheduler-title Auto-scheduler
 
   div.row
     form.form-inline ng-if="schedulerSolvable()"
