+++ features/step_definitions/autoscheduler_steps.rb
@@ -1,0 +1,11 @@
+Then(/^autoscheduler should(n't)? be visible$/) do |negate|
+  pending # express the regexp above with the code you wish you had
+end
+
+Then(/^I should(n't)? see "(.*?)" in sections nobody can teach$/) do |negate, section|
+  pending # express the regexp above with the code you wish you had
+end
+
+Then(/^I should(n't)? see "(.*?)" in GSIs who can't teach enough classes to fulfill their appointment$/) do |negate, name|
+  pending # express the regexp above with the code you wish you had
+end
